uuid pk

+++ src/infrastructure/model/page.rs
@@ -10,11 +10,11 @@ pub struct PageQuery {
 /// 页数据结果
 #[derive(Debug, Serialize)]
 pub struct PageResult<T> {
-    pub page: u8,
+    pub page: u64,
 
-    pub page_size: u8,
+    pub page_size: u64,
 
-    pub record_total: u8,
+    pub record_total: u64,
 
     pub records: Vec<T>,
 }
